refactor: fold slot resolver helper into slot_v1.go
The api package follows a one-file-per-endpoint convention; a standalone
slot_resolver.go didn't fit. Move resolveSlotOrHash next to APISlotV1 in
slot_v1.go since it's the natural shared slot file.
